代码路径风格改变。
