+++ index.html
@@ -4,8 +4,7 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
     <link rel="stylesheet" href="./src/assets/css/global.css">
-    <meta name="description" content="" />
-    <meta name="keywords" content="" />
+    <link rel="stylesheet" href="./src/assets/icon/iconfont.css"> <!--引入第三方图标库-->
     <title>labsys</title>
   </head>
   <body>
