+++ src/assets/icon/demo_index.html
@@ -37,6 +37,12 @@
               </li>
           
             <li class="dib">
+              <span class="icon iconfont">&#xe60a;</span>
+                <div class="name">登录头像</div>
+                <div class="code-name">&amp;#xe60a;</div>
+              </li>
+          
+            <li class="dib">
               <span class="icon iconfont">&#xe603;</span>
                 <div class="name">电脑</div>
                 <div class="code-name">&amp;#xe603;</div>
@@ -188,6 +194,15 @@
               电脑
             </div>
             <div class="code-name">.aliIcon-lab-diannao
+            </div>
+          </li>
+          
+          <li class="dib">
+            <span class="icon iconfont aliIcon-lab-denglutouxiang"></span>
+            <div class="name">
+              登录头像
+            </div>
+            <div class="code-name">.aliIcon-lab-denglutouxiang
             </div>
           </li>
           
@@ -370,6 +385,14 @@
                 </svg>
                 <div class="name">电脑</div>
                 <div class="code-name">#aliIcon-lab-diannao</div>
+            </li>
+          
+            <li class="dib">
+                <svg class="icon svg-icon" aria-hidden="true">
+                  <use xlink:href="#aliIcon-lab-denglutouxiang"></use>
+                </svg>
+                <div class="name">登录头像</div>
+                <div class="code-name">#aliIcon-lab-denglutouxiang</div>
             </li>
           
             <li class="dib">
