+++ src/assets/icon/iconfont.json
@@ -13,6 +13,13 @@
       "unicode_decimal": 58916
     },
     {
+      "icon_id": "4286391",
+      "name": "登录头像",
+      "font_class": "denglutouxiang",
+      "unicode": "e60a",
+      "unicode_decimal": 58890
+    },
+    {
       "icon_id": "4824990",
       "name": "电脑",
       "font_class": "diannao2",
