+++ src/module/door/router/index.js
@@ -1,0 +1,19 @@
+// import Home from '../components/home/Home'
+import Door from '../page/Door'
+import pagePreloading from '../components/pagePreloading'
+import headMenu from '../components/headMenu'
+import Login01 from '../page/Login01'
+
+
+export default [{
+  path: '/',
+  component: Door,
+  name: '营销管理',
+  hidden: false,
+  children:[
+    {path: '/door/pagepreloading', component: pagePreloading, hidden: false},
+    {path: '/door/headmenu', component: headMenu, hidden: false},
+    {path: '/door/login', component: Login01, hidden: false},
+
+  ]}
+]
